201711051834 fix the code statments execution order
putting debug_tracer.attribute_ before each attribute/property assignment.

to add the it to the log in it's correct execution order becuase if  another object is assigned to this attribute it's gonna be write the object debug_tracer. lines/data before the assigned attribute itself in the execution order.

wrong order:
>>>>>>>> call Field.value()
	@      @ self.__value : 11
-------- exit ()Field.value
@      @ [_pk] : 11

correct order:
@      @ [_pk] : 11
>>>>>>>> call Field.value()
	@      @ self.__value : 11
-------- exit ()Field.value